Refactor authentication and authorization to make them combinable
The patch refactors AuthenticationHandler and AuthorizationHandler such
that the subject generated by the first is given to the second.

Both interfaces are cleaned up a bit. The specification of the
interfaces are updated to let the caller detect anonymous sessions.
dCache can use this to apply special mappings for anonymous sessions.

The authorization interface is modified to allow a plugin to return an
authorized Subject. This allows authorization plugins to apply custom
mappings or extract identifying information from an authorization token.