xrootd4j-gsi: Remove GlobusPrincipal from principal set
The GlobusPrincipal that was added was incorrectly formatted and
it used the DN of the proxy certififcate, not of the certificate
that issued the proxy certificate.

For now I simply removed the principal completely. The certificate
chain is available in the public credentials set. In dCache the
DN will be extracted by gPlazma and added as a GlobusPrincipal.