Add support for chunked replies.
The code is imported from dCache, but rewritten to follow the
structure of Netty's ChunkedWriteHandler and generalize to allow
other chunked replies that ReadResponse.

We cannot use Netty's ChunkedWriteHandler directly as we would
like to support out of order delivery of xrootd replies and
also do a limited form of read-ahead.

Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/5146/