Reimplement ChunkedResponseWriteHandler based on Netty 3.6.3
Our original ChunkedResponseWriteHandler was based on an older version
of Netty (probably 3.2.4). Since then the Netty project has fixed
plenty of bugs in their ChunkedResponseHandler on which we based
our handler.

This patch reimplements ChunkedResponseWriteHandler based on the
latest version in Netty 3.6.3.

I dropped the read-ahead logic as their was no evidence that it made
a difference. It was just an additional buffer on top of the socket
send buffer and thus seemed somewhat pointless.

Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/5405/