xrootd4j: Null terminate several strings
I observed that xrdfs is behaving non-deterministically and suspect it
expects the response to kXR_locate to be null terminated. The spec doesn't
state whether the string is null terminated or not, but the non-determinstic
behaviour goes away when it is null terminated.

After careful reading of the spec, I concluded that then open and login
strings should use null terminated strings too.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/6648/