xrootd4j: Fix race condition in authentication
The authentication filter suffered from races if used with an unordered
executor (as it the case in dCache). This is a potential security problem, as a
check whether a request could be allowed through even when not authenticated.

The patch also fixes propagation of a disconnect channel event, and injects the
current session into more requests.

Target: trunk, 1.3
Require-notes: yes
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/7440/