Update to Netty 4
Netty 4 introduces major API changes, including:

- All classes are under the io.netty package.
- ChannelBuffer was renamed to ByteBuf.
- Buffers are reference counted and optionally pooled.
- ChannelEvent objects have been replaced by direct method
  invocation.
- Handler interface and base classes have been renamed.
- The thread model has changed completely.

Many of the changes are easily applied. Reference counting buffers
requires careful design of handlers, although for many xrootd
handlers most details are taken care of by the XrootdRequestHandler
abstract base class.

The other major change is the new thread model. The ordered and
memory limiting executor handler is gone. Some support for
alternate executors is present in the new pipeline implementation,
but there is no replacement for the memory limiting aspects.
For the standalone server this is hardly an issue, as the gains
of using the separate executor were probably non-existing. For
dCache, we have to manually hand blocking operations over to an
executor, independently of Netty.

I changed the design of the message decoding and buffer copy in
the decoder. The message implementation is now responsible for
copying the buffer if needed (only the read messages need it).
It avoids unnecessary copies and simplifies reference counting.

Due to the massive API changes, channel handler plugins have to
be updated to work with the new version. Therefore I bumped the
version of xrootd4j to 2.0.

Target: master
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/7545/