xrootd4j: Fix computed length of bucket encoding
Motivation:

The xrootd authentication protocol uses a flexible encoding scheme of buckets.
When serializing such buckets, our calculated length of the nested bucket type
fails to include the length of the header and the terminal fields. As a result
the GSI module fails.

The failure is limited to xrootd4j 2.1 as earlier versions only used the
calculated length for preallocation.

Modification:

Add the headers and terminal fields to the lenght.

Result:

Fixes GSI authentication.

Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8365/
Target: master, 2.1