xrootd4j: Expose response details in asynchronous response
Motivation:

The xrootd protocol allows asynhronous responses. In this case the server
first replies synchronous that it will deliver an asynchronous response and
later it wraps the real response in a asynresp reply.

Modification:

Exposes the wrapped response and the status code of the response. This allows
other handlers to access the response details.

Result:

Better logging in dCache and enables the redirector plugin to be fixed.

Target: master, 3.0, 2.1
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8827/