xrootd4j: Enable HAProxy aware authorization
Authorization plugins can authorize by local and remote socket address. E.g.
the alice authorization token plugin verifies that the token is for a TURL for
the destination address. In the pressence of an HAProxy in front of the server,
the destination address is different from the local address of the channel and
thus the alice token plugin fails.

This patch injects the actual source and destination address into the
authorization plugin.

Target: master,3.2
Acked-by: Anupam Ashish <anupam.ashish@desy.de>

Reviewed at https://rb.dcache.org/r/9832/