xrootd4j: distinguish correctly between kXR_wait and kXR_waitresp
Motivation:

The xrootd protocol specifies that the server can respond to the client
with instructions to wait.  There are two such responses, 4005 and 4006;
the former simply means wait, then retry.  The latter, however, means
wait for an asynchronous "go ahead" from the server.  In this case,
arriving at the max timeout with no instruction from the server
constitutes a failure, and should not trigger a retry of the
request which was checked by the wait.

Unfortunately, these two distinct instructions from the server
are currently garbled in the TPC handlers.  In particular,
the client, upon receiving 4006, waits for ten seconds, then
retries.  This provokes a "duplicate request" failure on the
server side.

Modification:

Handle 4005 as before, but distinguish this from 4006 by
timing out and failing.  The code to act on the server's
asynchronous 'attention' response (kXR_attn) is already
in place.

Result:

Duplicate request failures no longer occur.

Target: master
Request: 3.3
Acked-by: Tigran