xrootd4j: add support for signed hash verification to server-side security
Motivation:

Third-party copy already supports (with one caveat, see below)
signed hash verification (sigver) when communicating with
a vanilla xrootd server.   We now wish to complete the support
by making the server side interactions also support sigver.

Modification:

All inbound requests to the server now carry their appropriate
security level settings.

The protocol response now conveys security level and forced
hash signing (for authentication which does not encrypt) to
the client.

The necessary sigver messages and handler have been implemented.

The necessary additions to the incoming request decoder
have been made.

In order to get signed hashes on the pool connection, the
pool login response must indicate "&P=unix" to the
client.  This also means that kXR_secOFrce must always
be set in order for signed hashes (unencrypted) to precede
the requests sent to the pool (depending on the security
level that has been set).  This change will be made in
the dcache-xrootd patch which follows.

As a consequence, TPC must now also support the unix
protocol when dcache to dcache third-party copying is
involved.  A patch adding the unix module follows this
one.

Result:

Signed hashes (encrypted/gsi on the door, unencrypted/unix
on the pool) are now supported, with the exception of
dcache-to-dcache TPC, which will be added separately.

Target: master
Acked-by: Tigran