xrootd4j: restructure client pipeline error handling
Motivation:

Code should avoid constructing an exception and then
passing it off to an exception handler.  The exception
should be thrown and caught.

Modification:

Change the 'doOn...' methods in the handlers to
throw an XrootdException.

Change the requestReceived method to place calls
to those handlers inside a try...catch block.

Have the catch block then pass the exception to
the exceptionCaught method, to ensure that
the client is properly closed, so that the
caller gets a response from the source server.

Result:

Code which conforms to best practice.

Target: master
Request: 3.3
Acked-by: Tigran