xrootd4j: correctly handle multiple authn protocols as indicated by server
Motivation:

While not clearly indicated in the xrootd protocol documentation,
it is logically possible for the server to indicate in its login
response a list of authentication protocols to be tried, one
after the other.  These are indicated in the 'sec' data by
successive '&P=' segments.

Currently we are not handling these correctly.  The GSI module
in particular assumes only one protocol will be defined;
the TPC client also loads all such modules in the order
specified in the dCache configuration, so they will sit
in the pipeline in that sequence, not in the sequence
required by the server.

Modification:

Move the parsing to a special class to hold this information.
Process this information into a list and a map (for easy
reference) in the LoginResponse.

Have the connect handler access this information and add
the necessary handlers to the pipeline.  These handlers,
constructed from the plugin factories,
are now stored in the client in a map until they are
needed.

The authentication handlers also need to behave differently
on exceptions.  Instead of failing fast, they need to
pass the original login response downstream to the
next handler.

Result:

No current change in behavior, since we only support GSI,
but we avoid potential parsing issues should servers send
back multiple protocol indications.

Target: master
Request: 3.3
Acked-by: Tigran