xrootd4j: redesign outbound TPC sigver handling
Motivation:

The current implementation of outbound third-party
support of signed hash requests adds a handler to
the client, which is then checked by the outbound
encoder to see whether the request being written
needs to be preceded by the signed hash request.

This slightly breaks the pipeline semantics of Netty.
A superior way of handling this is to add another
encoder (in penultimate position) to the pipeline
rather than to the client, and to delegate responsibility
for generating the sigver request to it.

Modification:

Removed generic marker and reimplemented the sigver
handler as an encoder.

Security handlers, which are responsible for providing
the encoder with an ecryption mechanism, add the
encoder to the pipeline just after the main outbound
encoder (so far, only the GSIClientAuthnHandler needs
to do so).

Result:

Code conforms more closely with Netty design.

Target: master
Acked-by: Tigran