xrootd4j: prevent NPE when constructing error response
Motivation:

When constructing an ErrorResponse, avoid a null message.
This is done in two places where the constructor is called,
but not in the third.   When data length is called, a null
message causes an NPE.

Modification:

Make the constructor do Strings.nullToEmpty on the parameter.

Result:

No NPE, and future calls to the constructor
need not explicitly call the Strings.nullToEmpty.

Target: master
Request: 3.3
Acked-by: Paul