xrootd4j: add ERROR status to tpc info
Motivation:

We need a way of invalidating the tpc info without
actually discarding it or evicting it from a cache.

Modfication:

Add Status.ERROR.  If verify() is called on this
info, it immediately returns this error status.

Result:

A destination server can now know, when it tries
to call open, for instance, that the client
open failed (so we do not rely on the client
to close the server connection in time).

Target: master
Request: 3.3
Acked-by: Tigran