pom: bump copyright year to 2019
Motivation:

The license-maven-plugin requires copyright year range to end with the
year of the last commit.  When tagging a release, further commits are
made.  This requires that all the pom.xml files bear the current year.

Modification:

Update copyright year range for the pom.xml files to the current year.

Result:

Ability to tag releases

Target: master
Request: 3.3