xrootd4j: add framework for decrypting and verifying signed hashes
Motivation:

Add server-side support for signed hash verification.

Second patch of four.

Modification:

Create a decoder for intercepting the buffer and verifying it
against the signed hash sent before the actual message.
Also factored out some shared code into an abstract class
for the decoders.

Swap in the decoder on the pipeline on successful authentication.

Modify the DHEncrypter to do decryption as well; allow
authentication handlers to return the appropriate decrypter.

Add setter method on the XrootdAuthenticationHandler so
that the server-wide signing policy can be injected.

Add security level information to the ProtocolResponse.

Implement the SigverRequest.

Result:

Framework for server-side verification is in place.

Target: master
Acked-by: Tigran