xrootd4j: improve client error messages
Motivation:

Details concerning stream and channel id, and other server internals,
should not appear in the error messages sent back to the client.

Modification:

Separate out error logging strictly for the purposes of the server
from the message included in the exception returned to the calling
client.

Result:

User does not see details that have no pertinence.

Target: master
Acked-by: Paul
Acked-by: Tigran