xrootd4j: remove rather than get from list of SecurityInfo
Motivation:

Code was originally written to observe the empty list semantics
for security info on the login response as indicating that
all the necessary handlers had been added to the pipeline
but none of them successfully authenticated the client.

Unfortunately, the 'remove' (instead of 'get') of each
protocol definition from the map was evidently not implemented.

Modification:

Change 'get' to 'remove'.

Result:

Success/failure checking is now correct.

Target: master
Acked-by: Tigran