xrootd4j: trim protocol string when there is no metdata
Motivation:

Patch #11457 (master@f9b85c02f3f94a322ef315790d417b5108368f65) refactored the SecurityInfo class.
This also introduced a small regression into the parsing of the information.

If the definition is a single or trailing "&P=" without and "," following, it evidently
needs to be trimmed.  Otherwise, the name does not map and the corresponding handler
is not added to the pipeline (in particular:  "unix@" vs "unix").

Modification:

Trim the string in this case.

Result:

Unix support works again.

Target: master
Acked-by: Tigran