xrootd4j-gsi: refactor request handling in preparation for 4.9/delegation support
Motivation:

The 4.9 version of the xrootd protocol introduces proxy delegation support,
and in the process also alters the protocol message passing structure and
procedure.

In order to support the updated protocol, the dCache library needs
to be refactored.  The second modification is aimed at isolating
the parts of the request handling which are subject to change from
version pre 4.9 to 4.9.

Modification:

    Create a new tree of abstract classes for GSI request handling.

    Implement the pre49 client and server request handlers based on
    the current GSIAuthenticationHandler and GSIClientAuthenticationHandler,
    leaving only the "switching" code in the latter.

    Compose the new classes into the GSIAuthenticationHandler and
    GSIClientAuthenticationHandler.

    Update the current classes to reference the composed handler
    instead of superclass constants and methods.

Result:

No visible changes for user.

Target: master
Acked-by: Tigran
Requires-notes: no
Requires-book: no