xrootd4j: do not reject 4.9 client
Motivation:

In preparation for supporting two different protocols,
the GSI handler checks versions and tries to match.

However, the code currently rejects the 4.9 client
saying it does not support it.   This is not necessary
and has caused our WLCG-DOMA deployment to be incompatible
with the testbed client.

Modification:

Accept 10400 client version and just use the current
client protocol.  The 10400 client should be
backward compatible.

Result:

Testbed should work again.

Target: master
Request: 3.4
Acked-by:  Paul