xrootd4j: change credential store API to proxy delegation API
Motivation:

Do not require storage and sharing of delegated proxies.
This simplifies the semantics and handling, and also
follows what the SLAC implementation does more closely.

Modification:

Changes to the API so that the end of the signing request
exchange with client results in the delivery rather
than storage of the full delegated credential.

This is passed along in the XrootdSession object
to the successive open request.

Result:

Cleaner, simpler proxy delegation.

NOTE that most of this patch simply involves renaming.

Target: master
Acked-by: Tigran