xrootd4j: fix OutboundProtocolRequest header
Motivation:

In order for the xrootd client to know whether to honor
hash signing, the xrootd server must return security level
information to the client in the ProtocolResponse.

The server will do this only if the following conditions
pertain:

- the server's own version is >= 3.1.0
- the protocol version supported by the client is >= 3.1.0
- the client sets the option in the ProtocolRequest to 'kXR_secreqs'

Currently, there are two defects in the TPC client code
preventing the server from conveying its security level to
dCache during TPC with dCache as destination:  the
protocol version is out of date, and the byte ordering
of the ProtocolRequest is incorrect.

As a result, if the security level of the dCache server
is > 0, the transfer fails because the TPC client does
not honor hash signing.

Modification:

1.  Update the protocol version.
2.  Fix the byte order in the OutboundProtocolRequest

Result:

With `sec.level all pedantic` set on the xrootd server,
dCache now honors hash signing and TPC with dCache
as destination succeeds.

Target: master
Request: 3.5
Requires-book: no
Requires-notes: yes
Patch: https://rb.dcache.org/r/11925/
Acked-by:  Olufemi