xrootd4j: update and regularize use of error codes throughout library
Motivation:

Protocol 4.0 adds a few more error codes
which need to be incorporated.

But in general, the libary was in need
(as is the dcache-module) of review in terms
of its use of error codes.

Often, generic errors could be replaced by
more specific ones; but kXR_ServerError
should also not be abused, especially
in light of the fact that the client
may decide to retry when it receives
this error.  There were also instances
when the GS (security) errors subcodes
were more appropriate.

Modification:

Reviewed and adjusted as needed.

Result:

Hopefully a more meaningful response
to the XrootD client is returned.

Target: master
Request: 3.5
Patch: https://rb.dcache.org/r/11957
Acked-by: Tigran
Acked-by: Lea