xrootd4j: do not swallow opaque data when parsing the QueryRequest
Motivation:

The QueryRequest embraces a number of different
subtypes, two of which are actually Path queries.
As such, opaque data is possibly sent with
them.  In the case of the checksum query,
this opaque data actually specifies the
type of checksum requested.

Currently, this opaque data is being
swallowed when processed by the (NOOP)
authorization handler.

Modfication:

Put the parsing where it belongs
(inside the request) and expose
getters for path and opaque,
to be used where appropriate.

Modify the ChecksumInfo
constructor so the parsing
need not be redone.

Result:

The checksum type is now
available where before
it had disappeared.

Partial solution to GitHub issue https://github.com/dCache/dcache/issues/5147

Target: master
Request: 3.5
Request: 3.4
Request: 3.3
Patch: https://rb.dcache.org/r/12018
Bug:  dCache/dCache#5147
Acked-by: Paul