xrootd4j: make AuthorizationHandler aware of ChannelHandlerContext
Motivation:

In preparation for supporting SciTokens in dCache's
xrootd door, a minor adjustments to the
construction of the Authorization handler
needsto be made.

The current API assumes such handling is isolated
from the Netty pipeline. Unfortunately, this
was somewhat short-sighted.

Modification:

Allow the ChannelHandlerContext to be
communicated to the AuthorizationHandler
via parameter on the factory method.

Result:

The authorization handler is no longer isolated
from Netty, which is essential for the
implementation of the SciToken handler.

Target: master
Patch: https://rb.dcache.org/r/12168/
Acked-by: Tigran