xrootd4j: catch errors thrown during bootload connect of the TPC client
Motivation:

If an error occurs (such as invalid source address)
during the connect sequence for the TPC client, it
gets propagated by being rethrown.  It seems that
certain errors are not passed to registered listeners
and appear as internal failures.

Modification:

Try to handle the error so as to provide, when possible,
a more meaningful error message to the client.

Result:

Failure with a better change of being understood.

Target: master
Request: 3.5
Request: 3.4
Acked-by: Tigran
Patch: https://rb.dcache.org/r/12232