xrootd4j: add property to enforce delegation only on the TPC client
Motivation:

The first versions of xrootd TPC allowed for
the generation of a proxy from the host cert and key,
or for the use of an externally refreshed credential.

As of 5.2 (xrootd 4.9), however, delegation of the
proxy is the usual scenario.   When we introduced
this, we also allowed the TPC client to fall back
to the previous method for backward compatibility.

This fallback, however, in most cases will probably
not be successful, as the endpoint configuration
needs to provide for accepting those credentials.

It makes more sense to have the default behavior
rely exclusively on delegation, and only allow
fallback if specifically configured for it.

Modification:

Add a property which defaults to true, and
only load the pre49 client authentication
module if false.

Result:

Less of a chance of a mysterious authentication
or authorization failure because of dynamic
adaptation which was not foreseen.

Target: master
Patch: https://rb.dcache.org/r/12248/
Acked-by: Tigran