xrootd4j: add support for TLS
Motivation:

TLS is a necessary precondition for SciToken
authorization, and a desideratum (at least in
some cases) for WAN data protection.

Modification:

Followed the protocol xrootd specifications
(5.0, tentative) in terms of the following:

- providing for optional vs strict (i.e,
     supporting both or only one of xroot, xroots)
- providing for both client and server initiation
- providing for the login, session and tpc options
- signed hash verification automatically turned off
     if TLS is active

Some refactoring was done in order to maintain
code symmetry (moving some functionality out of
dCache proper and into the library –– in particular,
the ProtocolRequest handling).

Most of the new code is in two larger classes:

The ServerProtocolFlags stores both the static
host configuration, but is also copied
on a per-session basis to record the dynamically
determined session settings;

TLSSessionInfo is the TLS control-center, so
so to speak, and is meant to be session-local.
This is also the place where the Netty SSLHandler
is initialized and added to the pipeline.

NOTE:  there is one feature in the
xrootd protocol which has been implemented
here provisionally in advance of its
inclusion:  --tps tls / tps.tls=1 CGI.
This is meant to override the server-side
TPC client's configuration, telling it that
it MUST use TLS (xroots).

Result:

TLS is provided by the library.

A patch for the dcache-xroot module enabling
TLS follows.

Target: master
Patch: https://rb.dcache.org/r/12257/
Acked-by: Tigran