xrootd4j: modifications to support xrootd.tls tpc and tpc.spr=xroots
Motivation:

In the initial implementation of TLS, the "requires TPC" option
was left out of the server protocol options because it was
unclear how this was to work.

This has now been clarified as follows:  "requires TPC" (the
equivalent of xrootd.tls tpc) means that the destination
server will block any clients which do not support TLS
(i.e., pre-version-5) from requesting TPC.  The TPC
requirement, however, is not sufficient to activate
TLS on the third-party connection.  That is determined
by two factors:  (a) the protocol expressed by the
client in connecting to the source server (xroots vs
xroot), and (b) what the source server actually
requires.   If (a) is 'xroots' and both source
and destination have TLS, it is used; else
the TPC should fail.  If it is 'xroot', then
the source server determines the TLS requirement.

Modification:

The flag is added to the server protocol.  A small
modification in how the requirement values are
verified allows distinguishing between whether
TPC is set and whether the server actually
requires TLS.  The TPC flag is checked in
the appropriate places against the incoming
client flags.

tpc.spr (source protocol) is now accessed
to determine whether the TPC client should
advertise itself as "wantTls" or "ableTls".

Some debugging statments have been
added and some trace-level promoted to
debug-level.

Result:

We now should behave like the vanilla
xrootd implementation with regard to
enforcement of TLS on the TPC connection.

Target: master
Patch: https://rb.dcache.org/r/12305
Request: 4.0.0
Acked-by: Tigran