xrootd4j: move static methods into bucket utils class
Motivation:

Preparation for adding a new protocol module, ZTN, which
has nothing to do with GSI.

There are legacy request, response and utilities classes
in xrootd4j which are instrumented with GSI authentication
in mind.  This logic does not belong in the generic xrootd4j
classes.

Modification:

This is the first of seven patches which will incrementally
refactor all GSI logic into separate classes and then
move them all into the xrootd4j-gsi module.

Here we extract static methods for deserialization of
"buckets" (which is an xrootd GSI protocol API/data structure)
out of the generic classes and place them in a utilities
class.

(The static references to the utilities class will be removed
in downstream patches).

Result:

No visible changes for the user.
Patch 1/7

Target: master
Patch: https://rb.dcache.org/r/12875/
Acked-by: Dmitry
Acked-by: Paul