xrootd4j: refactor AuthenticationRequest to use BucketData
Motivation:

Preparation for adding a new protocol module, ZTN, which
has nothing to do with GSI.

There are legacy request, response and utilities classes
in xrootd4j which are instrumented with GSI authentication
in mind.  This logic does not belong in the generic xrootd4j
classes.

Modification:

This patch abstracts Authentication Request bucket data
into a struct-like object and places it in the utils class.

It also modifies the handlers in the GSI module to take
the raw credential bytes and deserialize them there.

The internal API for the handlers now reflects the additional
parameter (Bucket Data) which is specific to GSI and not
a part of the Authentication Request per se.

Result:

No visibile changes for user.
Patch 3/7.

Target: master
Patch: https://rb.dcache.org/r/12877/
Acked-by: Dmitry