xrootd4j: refactor outbound authentication request
Motivation:

Preparation for adding a new protocol module, ZTN, which
has nothing to do with GSI.

There are legacy request, response and utilities classes
in xrootd4j which are instrumented with GSI authentication
in mind.  This logic does not belong in the generic xrootd4j
classes.

Modification:

We given the outbound request (used by the TPC client)
a Consumer/Function which takes a byte buffer and
serializes it.

Length computation specific to GSI is moved out
of the request object.

Result:

No visible changes for user.
Patch 4/7.

Target: master
Patch: https://rb.dcache.org/r/12878/
Acked-by: Dmitry