xrootd4j: refactor AuthenticationResponse
Motivation:

Preparation for adding a new protocol module, ZTN, which
has nothing to do with GSI.

There are legacy request, response and utilities classes
in xrootd4j which are instrumented with GSI authentication
in mind.  This logic does not belong in the generic xrootd4j
classes.

Modfication:

Similar to the outbound request, we provide the authentication
response with a consumer capable of serializing the raw
bytes into the correct data type.

Again, GSI-specific length computation is moved out of the
response object.

Result:

No visibile changes for the user.
Patch 5/7.

Target:  master
Patch: https://rb.dcache.org/r/12879/
Acked-by: Tigran