xrootd4j: send correct info on UNIX protocol
Motivation:

The UNIX module is provided for our TPC
client to be able to obey the protocol
directive sent from our pools when signed hash
verification is on and force signing is set
to true.  This bizarre behavior is
for compatibility with other xrootd clients,
which will not observe hash signing unless
some authn protocol is enforced.

The SLAC unix protocol specifies that the
"credential" part of the authn request should
consist of a space-separated user uid + gid.
Currently we are not following protocol
because our message consists of the
username.   This has not provoked
failures, however, because only we
require this and the pool does not
check credential contents as it is unnecessary,
given the UUID necessary to open the mover.

However, for correctness' sake, the
protocol data should conform to the
specification.

Modification:

The tpc info object is given uid and gid
fields which can be set, and the unix
handler serializes the credential using
these instead of user name. In the absence
of a uid or gid, we fall back to the username.

Result:

We now follow xroot specification for
the UNIX protocol.

NOTE that a corresponding change in
dCache will need to pull the uid and gid
out of the Subject on the pool and
provide them to the XrootdTpcInfo object.

Target: master
Patch: https://rb.dcache.org/r/12895/
Acked-by: Dmitry