xrootd4j: move and rename deserializeProtocol method
Motivation:

The static method for reading the four-byte
protocol from the ByteBuf in the GSI utilities method
is actually generalizable and will also be reused by
the ZTN protocol, so it should be made visible outside
the GSI module.

Modification:

Move it to the decoder (as we did a similar write
method to the encoder) and rename it.

Also turn the string prefixes and max length
values into constants that are reusable.

Result:

Method can be reused outside of GSI.
No changes visible to user.

Target: master
Patch: https://rb.dcache.org/r/12905
Acked-by: Dmitry